add missing int -> float cast for pass priority in effect files
Former-commit-id: d09bcd57fd4ade854c3745c9934bc3619578a2c9

+++ framework/src/minko/file/EffectParser.cpp
@@ -281,7 +281,7 @@ EffectParser::parsePasses(const Json::Value&		root,
 			_passIncludes[passCopy] = _passIncludes[pass];
 			_shaderIncludes[passCopy->program()->vertexShader()] = _shaderIncludes[pass->program()->vertexShader()];
 			_shaderIncludes[passCopy->program()->fragmentShader()] = _shaderIncludes[pass->program()->fragmentShader()];
-			passCopy->states()->priority(passesValue.size() - passId);
+			passCopy->states()->priority((float)(passesValue.size() - passId));
 
 			passes.push_back(passCopy);
